--- HW2/Makefile
@@ -1,20 +1,0 @@
-CC = gcc
-CFLAGS = -Wall -Werror
-
-SERVER_SRC = hw2_receiver.c
-CLIENT_SRC = hw2_sender.c
-
-SERVER_TARGET = receiver
-CLIENT_TARGET = sender
-
-all: $(SERVER_TARGET) $(CLIENT_TARGET)
-
-$(SERVER_TARGET): $(SERVER_SRC)
-	$(CC) $(CFLAGS) -o $(SERVER_TARGET) $(SERVER_SRC)
-
-$(CLIENT_TARGET): $(CLIENT_SRC)
-	$(CC) $(CFLAGS) -o $(CLIENT_TARGET) $(CLIENT_SRC)
-
-
-clean:
-	rm -f $(SERVER_TARGET) $(CLIENT_TARGET)
