+++ HW2_22000019/receiver
