+++ HW3_22000019/Makefile
@@ -1,0 +1,20 @@
+CC = gcc
+CFLAGS = -Wall -Werror
+
+SERVER_SRC = hw3_22000019_server.c
+CLIENT_SRC = hw3_22000019_client.c
+
+CLIENT_TARGET = client
+SERVER_TARGET = server
+
+all: $(SERVER_TARGET) $(CLIENT_TARGET)
+
+$(SERVER_TARGET): $(SERVER_SRC)
+	$(CC) $(CFLAGS) -o $(SERVER_TARGET) $(SERVER_SRC)
+
+$(CLIENT_TARGET): $(CLIENT_SRC)
+	$(CC) $(CFLAGS) -o $(CLIENT_TARGET) $(CLIENT_SRC)
+
+
+clean:
+	rm -f $(SERVER_TARGET) $(CLIENT_TARGET)
